Move hello to pkg in benchamrks folder
