+++ benchmarks/hello/hello.go
@@ -1,9 +1,9 @@
-package main
+package hello
 
 import (
 	"fmt"
 )
 
-func main() {
+func Hello() {
 	fmt.Println("Hello")
 }
