+++ benchmarks/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/estensen/runtime-systems/benchmarks/hello"
+
+func main() {
+	hello.Hello()
+}
